backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["PashaDem <dem-pasha@lenta.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
pydantic = "^1.10.7"
pyjwt = "^2.6.0"
azure-cosmos = "^4.3.1"
azure-identity = "^1.12.0"
aiohttp = "^3.8.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
